requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
asgiref==3.11.1
asttokens==3.0.1
blis==1.3.3
boto3==1.42.54
boto3-stubs==1.42.54
botocore==1.42.54
botocore-stubs==1.42.41
catalogue==2.0.10
certifi==2026.1.4
cffi==2.0.0
charset-normalizer==3.4.4
click==8.3.1
cloudpathlib==0.23.0
cloudpickle==3.1.2
comm==0.2.3
confection==0.1.5
contourpy==1.3.2
cryptography==46.0.5
cuda-bindings==12.9.4
cuda-pathfinder==1.3.4
cycler==0.12.1
cymem==2.0.13
debugpy==1.8.20
decorator==5.2.1
Django==5.2.11
django-emoji==2.2.2
duckdb==1.4.4
emoji==1.7.0
exceptiongroup==1.3.1
executing==2.2.1
filelock==3.24.3
fonttools==4.61.1
fsspec==2026.2.0
google-api-core==2.30.0
google-api-python-client==2.190.0
google-auth==2.48.0
google-auth-httplib2==0.3.0
google-auth-oauthlib==1.2.4
googleapis-common-protos==1.72.0
httplib2==0.31.2
idna==3.11
ipykernel==6.31.0
ipython==8.38.0
jedi==0.19.2
Jinja2==3.1.6
jmespath==1.1.0
joblib==1.5.3
jupyter_client==8.8.0
jupyter_core==5.9.1
kiwisolver==1.4.9
langcodes==3.5.1
librt==0.8.1
llvmlite==0.46.0
markdown-it-py==4.0.0
MarkupSafe==3.0.3
matplotlib==3.10.8
matplotlib-inline==0.2.1
mdurl==0.1.2
mpmath==1.3.0
murmurhash==1.0.15
mypy==1.19.1
mypy_extensions==1.1.0
narwhals==2.17.0
nest-asyncio==1.6.0
networkx==3.4.2
numba==0.64.0
numpy==2.2.6
oauthlib==3.3.1
packaging==26.0
pandas==2.3.3
pandas-stubs==2.3.3.260113
parso==0.8.6
pathlib_abc==0.1.1
pathspec==1.0.4
pathy==0.11.0
pexpect==4.9.0
pillow==12.1.1
platformdirs==4.9.2
plotly==6.5.2
preshed==3.0.12
prompt_toolkit==3.0.52
proto-plus==1.27.1
protobuf==6.33.5
psutil==7.2.2
pt_core_news_lg @ https://github.com/explosion/spacy-models/releases/download/pt_core_news_lg-3.8.0/pt_core_news_lg-3.8.0-py3-none-any.whl#sha256=2561c9a72a938d37141e9694e1a36d25061a44ce7e4f3bad2d3fa3bb836191af
ptyprocess==0.7.0
pure_eval==0.2.3
pyasn1==0.6.2
pyasn1_modules==0.4.2
pycparser==3.0
pydantic==2.12.5
pydantic_core==2.41.5
Pygments==2.19.2
pynndescent==0.6.0
pyparsing==3.3.2
python-dateutil==2.9.0.post0
python-dotenv==1.2.1
pytz==2025.2
pyxdg==0.28
pyzmq==27.1.0
requests==2.32.5
requests-oauthlib==2.0.0
rich==14.3.3
rsa==4.9.1
s3fs==0.4.2
s3transfer==0.16.0
scikit-learn==1.7.2
scipy==1.15.3
seaborn==0.13.2
shellingham==1.5.4
six==1.17.0
smart-open==6.4.0
sns==0.1
spacy==3.8.11
spacy-legacy==3.0.12
spacy-loggers==1.0.5
spyder-kernels==3.1.3
sqlparse==0.5.5
srsly==2.5.2
stack-data==0.6.3
sympy==1.14.0
thinc==8.3.10
threadpoolctl==3.6.0
tomli==2.4.0
torch==2.10.0+cpu
torchvision==0.25.0+cpu
tornado==6.5.4
tqdm==4.67.3
traitlets==5.14.3
typer==0.24.1
typer-slim==0.24.0
types-awscrt==0.31.2
types-pytz==2025.2.0.20251108
types-s3transfer==0.16.0
typing-inspection==0.4.2
typing_extensions==4.15.0
tzdata==2025.3
umap-learn==0.5.11
uritemplate==4.2.0
urllib3==2.6.3
wasabi==1.1.3
wcwidth==0.6.0
weasel==0.4.3
wrapt==2.1.1
wurlitzer==3.1.1
